requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.1
requests